move normalizeAccountId and legacizeAccountId to common
