+++ packages/blockchain-utils-validation/package.json
@@ -25,6 +25,7 @@
   },
   "dependencies": {
     "@ceramicnetwork/blockchain-utils-linking": "^1.5.0-rc.0",
+    "@ceramicnetwork/common": "^1.8.0-rc.0",
     "@ethersproject/contracts": "^5.4.1",
     "@ethersproject/providers": "^5.4.5",
     "@ethersproject/wallet": "^5.4.0",
