+++ packages/blockchain-utils-validation/src/blockchains/tezos.ts
@@ -1,10 +1,10 @@
-import { AccountId, ChainId } from 'caip'
+import { ChainId } from 'caip'
 import { RemoteSigner } from '@taquito/remote-signer'
 import { BlockchainHandler } from '../blockchain-handler'
 import { LinkProof, tezos } from '@ceramicnetwork/blockchain-utils-linking'
 import fetch from 'cross-fetch'
 import * as uint8arrays from 'uint8arrays'
-import { normalizeAccountId } from '../util'
+import { normalizeAccountId } from '@ceramicnetwork/common'
 
 export const ADDRESS_NOT_FOUND_ERROR = new Error(`Address not found on the Tezos blockchain`)
 export const PUBLIC_KEY_NOT_PUBLISHED_ERROR = new Error(
