+++ packages/common/package.json
@@ -39,6 +39,7 @@
     "lodash.clonedeep": "^4.5.0",
     "logfmt": "^1.3.2",
     "rxjs": "^7.0.0",
+    "caip": "~1.0.0",
     "uint8arrays": "^2.0.5"
   },
   "devDependencies": {
