+++ packages/common/src/index.ts
@@ -6,6 +6,7 @@ export * from './stream'
 export * from './utils/http-utils'
 export * from './utils/stream-utils'
 export * from './utils/test-utils'
+export * from './utils/accountid-utils'
 export * from './logger-provider'
 export * from './loggers'
 export * from './networks'
