+++ packages/stream-caip10-link-handler/src/caip10-link-handler.ts
@@ -11,7 +11,7 @@ import {
   StreamState,
   StreamUtils,
 } from '@ceramicnetwork/common'
-import { normalizeAccountId } from './util'
+import { legacizeAccountId } from '@ceramicnetwork/common'
 
 export class Caip10LinkHandler implements StreamHandler<Caip10Link> {
   get type(): number {
@@ -113,7 +113,7 @@ export class Caip10LinkHandler implements StreamHandler<Caip10Link> {
     // TODO: handle CAIP-10 addresses in proof generation of 3id-blockchain-utils
     const account: string = validProof.account || validProof.address
 
-    let legacyAccountCaip10 = normalizeAccountId(account)
+    let legacyAccountCaip10 = legacizeAccountId(account)
 
     // We can assume that controllers will always follow the legacy CAIP-10 Link format
     let legacyControllerCaip10 = state.metadata.controllers[0]
